feat: Refactor AudioPlayer component and introduce AudioTrackPlayer for improved audio handling
